Move classes to the right packages
