+++ orchestrator/src/test/kotlin/com/ntgjvmagent/orchestrator/integration/agent/AgentControllerIT.kt
@@ -1,7 +1,7 @@
 package com.ntgjvmagent.orchestrator.integration.agent
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.ntgjvmagent.orchestrator.agent.web.dto.AgentRequest
+import com.ntgjvmagent.orchestrator.dto.AgentRequestDto
 import com.ntgjvmagent.orchestrator.integration.BaseIntegrationTest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
@@ -31,7 +31,7 @@ class AgentControllerIT
         @DisplayName("should create and retrieve agent successfully")
         fun shouldCreateAndFetchAgent() {
             val request =
-                AgentRequest(
+                AgentRequestDto(
                     name = "API Agent",
                     model = "gpt-4o-mini",
                     description = "Integration test agent",
