+++ orchestrator/src/test/kotlin/com/ntgjvmagent/orchestrator/integration/agent/AgentRepositoryIT.kt
@@ -1,8 +1,8 @@
 package com.ntgjvmagent.orchestrator.integration.agent
 
-import com.ntgjvmagent.orchestrator.agent.domain.Agent
-import com.ntgjvmagent.orchestrator.agent.domain.AgentRepository
+import com.ntgjvmagent.orchestrator.entity.Agent
 import com.ntgjvmagent.orchestrator.integration.BaseIntegrationTest
+import com.ntgjvmagent.orchestrator.repository.AgentRepository
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
